Added a command handler

+++ src/commands/moderation/purge.js
@@ -1,31 +1,10 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
-import { updateSlashCommands } from '../../lib/updateSlashCommands.js';
 import { logAction } from '../../lib/util/util.js';
-import { emotes, embedColors, prefix } from '../../lib/config/config.js';
+import { emotes, embedColors } from '../../lib/config/config.js';
+import { handle } from '../../lib/commandHandler.js';
 import * as embed from '../../lib/util/embeds.js';
 
-export async function main() {
-  const { client } = await import('../../bot.js');
-
-  // Listen for purge commands
-  client.on('messageCreate', async message => {
-    if (!message.content.startsWith(prefix) || message.author.bot) return;
-    const args = message.content.slice(1).trim().split(' ').filter(str => str !== '');
-    const command = args.shift().toLowerCase();
-    if (!(command == 'purge')) return;
-    if (!(message.member.permissions.has('MANAGE_MESSAGES'))) return message.react(emotes.error);
-    
-    const userId = await (async () => {
-      try { return message.mentions.users.first() === undefined && args[0] !== args.at(-1) ? args[1].replace(/[\\<>@#&!]/g, '') : message.mentions.users.first().id; } 
-      catch { return null; }
-    })();
-    if (userId !== null && !(userId.match(/^[0-9]{15,18}/))) return message.reply(await embed.punishmentFail('Invalid user.'));
-    const messageCount = args[0];
-    const moderator = message.author;
-
-    await purge(message, messageCount, userId, moderator);
-  });
-
+export async function main(client) {
   // Create purge slash command
   const purgeData = new SlashCommandBuilder()
     .setName('purge')
@@ -37,25 +16,9 @@ export async function main() {
       .setDescription('Enter a user to purge')
       .setRequired(false));
     
-
-  // Listen for purge interactions
-  client.on('interactionCreate', async interaction => {
-    if (!interaction.isCommand()) return;
-    if (!(interaction.commandName === 'purge')) return;
-    if (!(interaction.member.permissions.has('MANAGE_MESSAGES'))) return interaction.reply({ content: 'Insufficient Permissions', ephemeral: true });
-    
-    const userId = interaction.options.get('user') === null ? null : interaction.options.get('user').value;
-    const messageCount = interaction.options.get('message_count').value;
-    const moderator = interaction.author;
-    
-    await purge(interaction, messageCount, userId, moderator);
-  });
-
-  updateSlashCommands(purgeData, 'purge');
-
-  async function purge(action, messageCount, userId, moderator) {
+  async function purge({ action, messageCount, userId, moderator }) {
     if (messageCount > 100 || messageCount < 1 || messageCount === undefined) return action.reply(await embed.punishmentFail('Message count must be between 1 and 100.'));
-    
+    if (userId !== null && !(userId.match(/^[0-9]{15,18}/))) return action.reply(await embed.punishmentFail('Invalid User.'));
     const channel = action.channel;
     const messages = (await channel.messages
       .fetch({ limit: messageCount, before: action.id }))
@@ -63,6 +26,8 @@ export async function main() {
         if (userId === null) return msg; 
         return msg.author.id === userId;
       });
+
+    if (messages.size === 0) return action.reply(await embed.punishmentFail(`No messages found from <@${userId}> in specified range.`));
     await channel.bulkDelete(messages);
 
     action.reply(await embed.createReplyEmbed(
@@ -75,4 +40,11 @@ export async function main() {
       { name: 'Channel', value: `${channel}` },
     ], { mod: moderator });
   }
+
+  handle(client, {
+    aliases: ['purge'],
+    requiredPerms: 'MANAGE_MESSAGES',
+    prefixedData: purgeData,
+    callback: purge
+  });
 }
